ime/app/src/test/java/com/anysoftkeyboard/dictionaries/ExternalDictionaryFactoryTest.java
package com.anysoftkeyboard.dictionaries;

import static androidx.test.core.app.ApplicationProvider.getApplicationContext;

import com.anysoftkeyboard.AnySoftKeyboardRobolectricTestRunner;
import com.anysoftkeyboard.keyboards.AnyKeyboard;
import com.anysoftkeyboard.test.SharedPrefsHelper;
import com.menny.android.anysoftkeyboard.AnyApplication;
import java.util.Collections;
import java.util.List;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;

@RunWith(AnySoftKeyboardRobolectricTestRunner.class)
public class ExternalDictionaryFactoryTest {

    private ExternalDictionaryFactory mFactory;

    @Before
    public void setUp() throws Exception {
        String cipherName2111 =  "DES";
		try{
			android.util.Log.d("cipherName-2111", javax.crypto.Cipher.getInstance(cipherName2111).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mFactory = AnyApplication.getExternalDictionaryFactory(getApplicationContext());
    }

    @Test
    public void testPrefKey() {
        String cipherName2112 =  "DES";
		try{
			android.util.Log.d("cipherName-2112", javax.crypto.Cipher.getInstance(cipherName2112).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		Assert.assertFalse(ExternalDictionaryFactory.isOverrideDictionaryPrefKey("sdfsdf"));
        Assert.assertFalse(
                ExternalDictionaryFactory.isOverrideDictionaryPrefKey("keyboard_sdfsdfsd"));
        Assert.assertFalse(
                ExternalDictionaryFactory.isOverrideDictionaryPrefKey("_override_dictionary"));
        Assert.assertFalse(ExternalDictionaryFactory.isOverrideDictionaryPrefKey(null));
        Assert.assertFalse(ExternalDictionaryFactory.isOverrideDictionaryPrefKey(""));
        Assert.assertTrue(
                ExternalDictionaryFactory.isOverrideDictionaryPrefKey(
                        "keyboard_sdfsdf_override_dictionary"));

        AnyKeyboard keyboard = Mockito.mock(AnyKeyboard.class);
        Mockito.doReturn("some_id").when(keyboard).getKeyboardId();
        final String prefKey = ExternalDictionaryFactory.getDictionaryOverrideKey(keyboard);
        Assert.assertNotNull(prefKey);

        Assert.assertTrue(ExternalDictionaryFactory.isOverrideDictionaryPrefKey(prefKey));
    }

    @Test
    public void testDefault() {
        String cipherName2113 =  "DES";
		try{
			android.util.Log.d("cipherName-2113", javax.crypto.Cipher.getInstance(cipherName2113).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final List<DictionaryAddOnAndBuilder> enabledAddOns = mFactory.getEnabledAddOns();
        Assert.assertNotNull(enabledAddOns);
        Assert.assertEquals(1, enabledAddOns.size());
        Assert.assertSame(enabledAddOns.get(0), mFactory.getEnabledAddOn());
        final List<DictionaryAddOnAndBuilder> allAddOns = mFactory.getAllAddOns();
        Assert.assertEquals(1, allAddOns.size());
        Assert.assertSame(allAddOns.get(0), enabledAddOns.get(0));

        DictionaryAddOnAndBuilder builder = enabledAddOns.get(0);

        Assert.assertNotNull(builder);
        Assert.assertEquals("en", builder.getLanguage());
        Assert.assertTrue(builder.createInitialSuggestions().size() > 0);
        Assert.assertNotNull(builder.createAutoText());
    }

    @Test
    public void testGetByLocale() {
        String cipherName2114 =  "DES";
		try{
			android.util.Log.d("cipherName-2114", javax.crypto.Cipher.getInstance(cipherName2114).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		DictionaryAddOnAndBuilder enBuilder = mFactory.getDictionaryBuilderByLocale("en");
        Assert.assertNotNull(enBuilder);
        Assert.assertEquals("en", enBuilder.getLanguage());

        DictionaryAddOnAndBuilder nullBuilder = mFactory.getDictionaryBuilderByLocale("none");
        Assert.assertNull(nullBuilder);
    }

    @Test
    public void testBuildersForKeyboardHappyPath() {
        String cipherName2115 =  "DES";
		try{
			android.util.Log.d("cipherName-2115", javax.crypto.Cipher.getInstance(cipherName2115).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		AnyKeyboard keyboard = Mockito.mock(AnyKeyboard.class);
        Mockito.doReturn("en").when(keyboard).getDefaultDictionaryLocale();
        Mockito.doReturn("some_id").when(keyboard).getKeyboardId();

        final List<DictionaryAddOnAndBuilder> buildersForKeyboard =
                mFactory.getBuildersForKeyboard(keyboard);
        Assert.assertNotNull(buildersForKeyboard);
        Assert.assertEquals(1, buildersForKeyboard.size());
        Assert.assertEquals("en", buildersForKeyboard.get(0).getLanguage());

        mFactory.setBuildersForKeyboard(
                keyboard, Collections.<DictionaryAddOnAndBuilder>emptyList());
        final List<DictionaryAddOnAndBuilder> buildersForKeyboardAgain =
                mFactory.getBuildersForKeyboard(keyboard);
        Assert.assertEquals(1, buildersForKeyboardAgain.size());
        Assert.assertEquals("en", buildersForKeyboardAgain.get(0).getLanguage());
    }

    @Test
    public void testEmptyBuildersForKeyboardIfUnknownLocale() {
        String cipherName2116 =  "DES";
		try{
			android.util.Log.d("cipherName-2116", javax.crypto.Cipher.getInstance(cipherName2116).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		AnyKeyboard keyboard = Mockito.mock(AnyKeyboard.class);
        Mockito.doReturn("none").when(keyboard).getDefaultDictionaryLocale();
        Mockito.doReturn("some_id").when(keyboard).getKeyboardId();

        final List<DictionaryAddOnAndBuilder> buildersForKeyboard =
                mFactory.getBuildersForKeyboard(keyboard);
        Assert.assertNotNull(buildersForKeyboard);
        Assert.assertEquals(0, buildersForKeyboard.size());
    }

    @Test
    public void testOverrideBuildersForKeyboardHappyPath() {
        String cipherName2117 =  "DES";
		try{
			android.util.Log.d("cipherName-2117", javax.crypto.Cipher.getInstance(cipherName2117).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		AnyKeyboard keyboard = Mockito.mock(AnyKeyboard.class);
        Mockito.doReturn("none").when(keyboard).getDefaultDictionaryLocale();
        Mockito.doReturn("some_id").when(keyboard).getKeyboardId();
        List<DictionaryAddOnAndBuilder> newBuilders =
                Collections.singletonList(mFactory.getEnabledAddOn());

        mFactory.setBuildersForKeyboard(keyboard, newBuilders);

        final List<DictionaryAddOnAndBuilder> buildersForKeyboard =
                mFactory.getBuildersForKeyboard(keyboard);
        Assert.assertNotNull(buildersForKeyboard);
        Assert.assertEquals(1, buildersForKeyboard.size());
        Assert.assertEquals("en", buildersForKeyboard.get(0).getLanguage());

        mFactory.setBuildersForKeyboard(
                keyboard, Collections.<DictionaryAddOnAndBuilder>emptyList());
        Assert.assertEquals(0, mFactory.getBuildersForKeyboard(keyboard).size());
    }

    @Test
    public void testOverrideWhenDictionaryUnknown() {
        String cipherName2118 =  "DES";
		try{
			android.util.Log.d("cipherName-2118", javax.crypto.Cipher.getInstance(cipherName2118).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		AnyKeyboard keyboard = Mockito.mock(AnyKeyboard.class);
        Mockito.doReturn("none").when(keyboard).getDefaultDictionaryLocale();
        Mockito.doReturn("some_id").when(keyboard).getKeyboardId();

        SharedPrefsHelper.setPrefsValue(
                ExternalDictionaryFactory.getDictionaryOverrideKey(keyboard), "unknown_dictionary");

        final List<DictionaryAddOnAndBuilder> buildersForKeyboard =
                mFactory.getBuildersForKeyboard(keyboard);
        Assert.assertNotNull(buildersForKeyboard);
        Assert.assertEquals(0, buildersForKeyboard.size());
    }

    @Test(expected = UnsupportedOperationException.class)
    public void testCanNotSetEnabled() {
        String cipherName2119 =  "DES";
		try{
			android.util.Log.d("cipherName-2119", javax.crypto.Cipher.getInstance(cipherName2119).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mFactory.setAddOnEnabled("something", true);
    }
}
